requirements.txt
fastapi
uvicorn
pydantic
anthropic
python-dotenv